+++ src/main/java/guideme/internal/screen/NavigationToolbar.java
@@ -2,10 +2,11 @@ package guideme.internal.screen;
 
 import guideme.Guide;
 import guideme.PageAnchor;
+import guideme.internal.GuideMEClient;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.Consumer;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.AbstractWidget;
-import net.minecraft.client.gui.components.Button;
 import org.jetbrains.annotations.Nullable;
 
 public class NavigationToolbar {
@@ -17,16 +18,61 @@ public class NavigationToolbar {
     @Nullable
     private Runnable closeCallback;
 
-    private Button backButton;
-    private Button forwardButton;
+    private boolean canSearch;
 
-    private int left;
+    private final GuideIconButton backButton;
+    private final GuideIconButton forwardButton;
+    private final GuideIconButton toggleFullWidthButton;
+    private final GuideIconButton closeButton;
+    private final GuideIconButton searchButton;
+
+    private final List<GuideIconButton> buttons = new ArrayList<>();
 
     public NavigationToolbar(@Nullable Guide guide) {
         this.guide = guide;
+
+        searchButton = new GuideIconButton(
+                0,
+                0,
+                GuideIconButton.Role.SEARCH,
+                this::startSearch);
+        backButton = new GuideIconButton(
+                0,
+                0,
+                GuideIconButton.Role.BACK,
+                () -> GuideNavigation.navigateBack(guide));
+        forwardButton = new GuideIconButton(
+                0,
+                0,
+                GuideIconButton.Role.FORWARD,
+                () -> GuideNavigation.navigateForward(guide));
+        toggleFullWidthButton = new GuideIconButton(
+                0,
+                0,
+                GuideIconButton.Role.OPEN_FULL_WIDTH_VIEW,
+                this::toggleFullWidth);
+        closeButton = new GuideIconButton(
+                0,
+                0,
+                GuideIconButton.Role.CLOSE,
+                () -> {
+                    if (closeCallback != null) {
+                        closeCallback.run();
+                    }
+                });
+
+        update();
     }
 
     public void update() {
+        updateLayout();
+
+        if (GuideMEClient.instance().isFullWidthLayout()) {
+            toggleFullWidthButton.setRole(GuideIconButton.Role.CLOSE_FULL_WIDTH_VIEW);
+        } else {
+            toggleFullWidthButton.setRole(GuideIconButton.Role.OPEN_FULL_WIDTH_VIEW);
+        }
+
         if (guide != null) {
             var history = GlobalInMemoryHistory.get(guide);
             backButton.active = history.peekBack().isPresent();
@@ -34,59 +80,77 @@ public class NavigationToolbar {
         }
     }
 
-    public void addToScreen(Consumer<AbstractWidget> addWidget, int topEdge, int rightEdge) {
-        if (closeCallback != null) {
-            var closeButton = new GuideIconButton(
-                    rightEdge - GuideIconButton.WIDTH,
-                    topEdge,
-                    GuideIconButton.Role.CLOSE,
-                    closeCallback);
-            addWidget.accept(closeButton);
-            rightEdge = closeButton.getX() - GAP;
+    public void addToScreen(Consumer<AbstractWidget> addWidget) {
+        addWidget.accept(closeButton);
+        addWidget.accept(toggleFullWidthButton);
+        if (guide != null) {
+            addWidget.accept(forwardButton);
+            addWidget.accept(backButton);
+        }
+        if (canSearch) {
+            addWidget.accept(searchButton);
+        }
+    }
+
+    private void updateLayout() {
+        buttons.clear();
+
+        if (canSearch) {
+            buttons.add(searchButton);
         }
 
         if (guide != null) {
-            forwardButton = new GuideIconButton(
-                    rightEdge - GuideIconButton.WIDTH,
-                    topEdge,
-                    GuideIconButton.Role.FORWARD,
-                    () -> GuideNavigation.navigateForward(guide));
-            addWidget.accept(forwardButton);
-            rightEdge = forwardButton.getX() - GAP;
-
-            backButton = new GuideIconButton(
-                    rightEdge - GuideIconButton.WIDTH,
-                    topEdge,
-                    GuideIconButton.Role.BACK,
-                    () -> GuideNavigation.navigateBack(guide));
-            addWidget.accept(backButton);
-            rightEdge = backButton.getX() - GAP;
+            buttons.add(backButton);
+            buttons.add(forwardButton);
         }
 
-        var canSearch = !(Minecraft.getInstance().screen instanceof GuideSearchScreen);
-        if (canSearch) {
-            var searchButton = new GuideIconButton(
-                    rightEdge - GuideIconButton.WIDTH,
-                    topEdge,
-                    GuideIconButton.Role.SEARCH,
-                    this::startSearch);
-            addWidget.accept(searchButton);
-            rightEdge = searchButton.getX() - GAP;
-        }
+        buttons.add(toggleFullWidthButton);
 
-        left = rightEdge;
-        update();
+        if (closeCallback != null) {
+            buttons.add(closeButton);
+        }
+    }
+
+    private void toggleFullWidth() {
+        GuideMEClient.instance().setFullWidthLayout(!GuideMEClient.instance().isFullWidthLayout());
     }
 
     private void startSearch() {
         GuideNavigation.navigateTo(guide, PageAnchor.page(GuideSearchScreen.PAGE_ID));
     }
 
-    public int getLeft() {
-        return left;
-    }
-
     public void setCloseCallback(@Nullable Runnable closeCallback) {
         this.closeCallback = closeCallback;
+        update();
     }
+
+    public boolean isCanSearch() {
+        return canSearch;
+    }
+
+    public void setCanSearch(boolean canSearch) {
+        this.canSearch = canSearch;
+        update();
+    }
+
+    public int getWidth() {
+        int width = 0;
+        for (var button : buttons) {
+            width += button.getWidth() + GAP;
+        }
+        return width;
+    }
+
+    public int getHeight() {
+        return 16;
+    }
+
+    public void move(int x, int y) {
+        for (var button : buttons) {
+            button.setX(x);
+            button.setY(y);
+            x += button.getWidth() + GAP;
+        }
+    }
+
 }