+++ docs/docs/changelog.md
@@ -3,11 +3,14 @@
 
 ## 2.6.0
 
+- Change the default layout of guides to be a centered column, and add a toolbar button to toggle between 
+  full-width and centered-column layout.
 - Fix navigating between guide pages not appending to navigation history
 - Add configuration screen
 - Make scaling of Guide UI independent of UI scale for scales 1 and 3, where the uniform Minecraft font
   has severe rendering issues. This behavior can be disabled in the config screen (Adaptive Scaling).
 - Added support for blast furnace recipes
+- Do not show a navigation bar for guides that do not have any navigation items
 
 ## 2.5.1